+++ apps/client/src/app/pages/blog/2023/03/1000-stars-on-github/1000-stars-on-github-page.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  host: { class: 'page' },
+  selector: 'gf-1000-stars-on-github-page',
+  styleUrls: ['./1000-stars-on-github-page.scss'],
+  templateUrl: './1000-stars-on-github-page.html'
+})
+export class ThousandStarsOnGitHubPageComponent {}